Added ability to send reservation to different computers
-Added the Showclix icon to the Status window
-Made ProcessHandler use try statements instead of regurgitating Exceptions thrown
-Added option to send reservation to another computer
--NOTE: I expect there to be a few bugs I'll uncover and fix later on
--Requires ShowclixDownload (in my Github repository list)
--Added ApproveForm, ConnectionPanel, Connections, NetworkHandler, and ClientConnection
--Receives connections on Port 9243
--Added checkbox (default unselected) to Setup to activate
--Can theoretically handle as many connections as you need
--Approve/Deny window appears before sending any information over
--Pops up with a ConnectionManager window when activated
--New connections are added to the window
--Can terminate connections at any time
--Connections print out to their respective output windows in ConnectionManager as well as to the management tab in ConnectionManager

+++ src/showclixscanner/ProcessHandler.java
@@ -9,30 +9,40 @@ import java.io.*;
 public class ProcessHandler {
 
   private static final String TASKLIST = "tasklist";
-  private static final String KILL = "taskkill /IM ";
+  private static final String KILL = "taskkill /T /IM ";
 
-  public static boolean isProcessRunging(String serviceName) throws Exception {
-    Process p = Runtime.getRuntime().exec(TASKLIST);
-    BufferedReader reader = new BufferedReader(new InputStreamReader(p.getInputStream()));
-    String line;
-    while ((line = reader.readLine()) != null) {
-      line = line.toLowerCase();
-      if (line.contains(serviceName)) {
-        System.out.println(serviceName + " found -- killing process.");
-        return true;
+  public static boolean isProcessRunging(String serviceName) {
+    try {
+      Process p = Runtime.getRuntime().exec(TASKLIST);
+      BufferedReader reader = new BufferedReader(new InputStreamReader(p.getInputStream()));
+      String line;
+      while ((line = reader.readLine()) != null) {
+        line = line.toLowerCase();
+        if (line.contains(serviceName)) {
+          System.out.println(serviceName + " found -- killing process.");
+          return true;
+        }
       }
+    } catch (IOException e) {
+      System.out.println("Error reading processes!");
+      e.printStackTrace();
     }
     return false;
 
   }
 
-  public static void killProcess(String serviceName) throws Exception {
-    Runtime.getRuntime().exec(KILL + serviceName);
+  public static void killProcess(String serviceName) {
+    try {
+      Runtime.getRuntime().exec(KILL + serviceName);
+    } catch (IOException e) {
+      System.out.println("Unable to kill process " + serviceName);
+      e.printStackTrace();
+    }
   }
 
-  public static void killFirefox() throws Exception {
+  public static void killFirefox() {
     String processName = "firefox.exe";
-    while (isProcessRunging(processName)) {
+    if (isProcessRunging(processName)) {
       killProcess(processName);
     }
   }
